Parsing Xbee CSV data into JSON
Created a test script to parse the Xbee test data provided from cloud into the JSON file as specified in the server protocol. The file lacks a substantial timestamp and does not account for problems with parsing the CSV file itself, namely the extra columns and commas. More information will be needed before this script can be incorporated into the main code.